Move Android marketing screenshot tests
