+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/AppStateResetRule.kt
@@ -1,9 +1,6 @@
 package com.flashcardsopensourceapp.app
 
-import android.app.LocaleConfig
-import android.app.LocaleManager
 import android.content.Context
-import android.os.LocaleList
 import androidx.core.app.NotificationManagerCompat
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.platform.app.InstrumentationRegistry
@@ -24,107 +21,12 @@ private val testOnlyPreferenceNames: List<String> = listOf(
 )
 
 open class AppStateResetRule : ExternalResource() {
-    companion object {
-        private const val localeApplyTimeoutMillis: Long = 5_000L
-    }
-
     override fun before() {
-        applyConfiguredMarketingScreenshotLocaleOverride()
         resetAndroidTestAppState()
     }
 
     override fun after() {
         resetAndroidTestAppState()
-        clearConfiguredMarketingScreenshotLocaleOverride()
-    }
-
-    private fun applyConfiguredMarketingScreenshotLocaleOverride() {
-        val localeConfig = configuredMarketingScreenshotLocaleConfigOrNull() ?: return
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val localeManager = context.getSystemService(LocaleManager::class.java)
-            ?: throw IllegalStateException("LocaleManager was unavailable while applying screenshot locale.")
-        val localeList = LocaleList.forLanguageTags(localeConfig.appLocaleTag)
-
-        InstrumentationRegistry.getInstrumentation().runOnMainSync {
-            localeManager.overrideLocaleConfig = LocaleConfig(localeList)
-            localeManager.applicationLocales = localeList
-        }
-        waitForOverrideLocaleConfigState(
-            localeManager = localeManager,
-            expectedLanguageTags = localeConfig.appLocaleTag,
-            phase = "applying marketing screenshot override locale config '${localeConfig.localePrefix}'"
-        )
-        waitForLocaleState(
-            localeManager = localeManager,
-            expectedLanguageTags = localeConfig.appLocaleTag,
-            phase = "applying marketing screenshot locale '${localeConfig.localePrefix}'"
-        )
-    }
-
-    private fun clearConfiguredMarketingScreenshotLocaleOverride() {
-        val localeConfig = configuredMarketingScreenshotLocaleConfigOrNull() ?: return
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val localeManager = context.getSystemService(LocaleManager::class.java)
-            ?: throw IllegalStateException("LocaleManager was unavailable while clearing screenshot locale.")
-
-        InstrumentationRegistry.getInstrumentation().runOnMainSync {
-            localeManager.applicationLocales = LocaleList.getEmptyLocaleList()
-            localeManager.overrideLocaleConfig = null
-        }
-        waitForLocaleState(
-            localeManager = localeManager,
-            expectedLanguageTags = "",
-            phase = "clearing marketing screenshot locale '${localeConfig.localePrefix}'"
-        )
-        waitForOverrideLocaleConfigState(
-            localeManager = localeManager,
-            expectedLanguageTags = null,
-            phase = "clearing marketing screenshot override locale config '${localeConfig.localePrefix}'"
-        )
-    }
-
-    private fun waitForLocaleState(
-        localeManager: LocaleManager,
-        expectedLanguageTags: String,
-        phase: String
-    ) {
-        val startTimeMillis = System.currentTimeMillis()
-        do {
-            InstrumentationRegistry.getInstrumentation().waitForIdleSync()
-            if (localeManager.applicationLocales.toLanguageTags() == expectedLanguageTags) {
-                return
-            }
-        } while (System.currentTimeMillis() - startTimeMillis < localeApplyTimeoutMillis)
-
-        throw IllegalStateException(
-            "Timed out after $localeApplyTimeoutMillis ms while $phase. " +
-                "expectedLanguageTags='$expectedLanguageTags' actualLanguageTags='${localeManager.applicationLocales.toLanguageTags()}'."
-        )
-    }
-
-    private fun waitForOverrideLocaleConfigState(
-        localeManager: LocaleManager,
-        expectedLanguageTags: String?,
-        phase: String
-    ) {
-        val startTimeMillis = System.currentTimeMillis()
-        do {
-            InstrumentationRegistry.getInstrumentation().waitForIdleSync()
-            val actualLanguageTags = localeManager.overrideLocaleConfig
-                ?.supportedLocales
-                ?.toLanguageTags()
-            if (actualLanguageTags == expectedLanguageTags) {
-                return
-            }
-        } while (System.currentTimeMillis() - startTimeMillis < localeApplyTimeoutMillis)
-
-        val actualLanguageTags = localeManager.overrideLocaleConfig
-            ?.supportedLocales
-            ?.toLanguageTags()
-        throw IllegalStateException(
-            "Timed out after $localeApplyTimeoutMillis ms while $phase. " +
-                "expectedLanguageTags='$expectedLanguageTags' actualLanguageTags='$actualLanguageTags'."
-        )
     }
 }
